requirements.txt
aiogram==2.14.2
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
cycler==0.10.0
decorator==4.4.2
easyocr==1.3.2
idna==3.2
imageio==2.9.0
imutils==0.5.4
kiwisolver==1.3.1
matplotlib==3.4.2
multidict==5.1.0
networkx==2.5.1
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.18.0
python-telegram-bot==13.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
tifffile==2021.7.2
torch==1.9.0
torchvision==0.10.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
yarl==1.6.3
